fix gallery:
When I zoom in on a picture, the picture need to expand on all of the screen, not just on the inner frame.
In the list, the corner of the boarder of each image need to be sharper, like in the design.
I want to be able to do also zoom out to both image and PDF file.
The item list need to be smaller, check out the shapes in the mocap, it need to be look almost the same.
The zoom tool, that display the current zoom size, need to looks like the zoom in Zeplin.
With -/+ buttons and the zoom will be in percentage, this design need to be solid and increase and decrease by 25%.

The "Loading File…" need to be in the same font as in the mocap that you get from me.
I not always see the PDF icon, when he needs to be.